Update method definition of each in failure/base
to conform with changes made to the definition by tarcieri in
ce6dc794 and 44eb8899.

Before this change any failure backends that inherit from base
and don't implement their own self.each method will throw a
wrong number of arguments (4 for 2) error when navigating to the
failure tab in the web view.